feat(8): Update controllers
- Move some "comment" endpoints to the post controller since they're more naturally tied to the post actions
