+++ common/src/main/java/org/wcdevs/blog/core/common/comment/CommentService.java
@@ -9,7 +9,7 @@ import org.wcdevs.blog.core.persistence.comment.PartialCommentDto;
  * Provides services to handle the business logic concerning the {@link Comment}s data.
  */
 public interface CommentService {
-  CommentDto createComment(CommentDto commentDto);
+  CommentDto createComment(String postSlug, CommentDto commentDto);
 
   CommentDto getComment(String commentAnchor);
