+++ persistence/src/main/java/org/wcdevs/blog/core/persistence/comment/CommentDto.java
@@ -37,10 +37,6 @@ public class CommentDto {
 
   @ToString.Include
   @EqualsAndHashCode.Include
-  private String postSlug;
-
-  @ToString.Include
-  @EqualsAndHashCode.Include
   private String parentCommentAnchor;
 
   @NotNull
