+++ rest/src/main/java/org/wcdevs/blog/core/rest/post/PostController.java
@@ -16,7 +16,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.wcdevs.blog.core.common.comment.CommentService;
 import org.wcdevs.blog.core.common.post.PostService;
+import org.wcdevs.blog.core.persistence.comment.CommentDto;
 import org.wcdevs.blog.core.persistence.post.PartialPostDto;
 import org.wcdevs.blog.core.persistence.post.PostDto;
 
@@ -28,6 +30,7 @@ import org.wcdevs.blog.core.persistence.post.PostDto;
 @RequiredArgsConstructor
 public class PostController {
   private final PostService postService;
+  private final CommentService commentService;
 
   @GetMapping("/")
   public ResponseEntity<Collection<PostDto>> getPosts() {
@@ -64,5 +67,21 @@ public class PostController {
   @PreAuthorize("hasAnyRole('EDITOR', 'AUTHOR')")
   public void deletePost(@PathVariable String postSlug) {
     postService.deletePost(postSlug);
+  }
+
+  @PostMapping("/{postSlug}/comment")
+  public ResponseEntity<CommentDto> createComment(@PathVariable String postSlug,
+                                                  @Validated @RequestBody CommentDto dto) {
+    return new ResponseEntity<>(commentService.createComment(postSlug, dto), HttpStatus.CREATED);
+  }
+
+  @GetMapping("{postSlug}/comment/all")
+  public ResponseEntity<Collection<CommentDto>> getAllPostComments(@PathVariable String postSlug) {
+    return new ResponseEntity<>(commentService.getAllPostComments(postSlug), HttpStatus.OK);
+  }
+
+  @GetMapping("{postSlug}/comment/root")
+  public ResponseEntity<Collection<CommentDto>> getRootPostComments(@PathVariable String postSlug) {
+    return new ResponseEntity<>(commentService.getRootPostComments(postSlug), HttpStatus.OK);
   }
 }
