+++ rest/src/test/resources/sample-comments.json
@@ -1,6 +1,5 @@
 [
   {
-    "postSlug": "integer-in-pretium-turpis-2011460751",
     "parentCommentAnchor": "a-parent-comment-fusce-eu-sagittis-tortor-2011460723",
     "body": "A child comment lorem ipsum dolor sit amet. Nec rhoncus leo facilisis. Nam scelerisque nibh sed nisl pellentesque, vel aliquam orci laoreet.",
     "publishedBy": "julia",
@@ -8,7 +7,6 @@
     "lastUpdated": "2021-12-06 14:57:49"
   },
   {
-    "postSlug": "integer-in-pretium-turpis-2011460751",
     "parentCommentAnchor": null,
     "body": "A parent comment fusce eu sagittis tortor. Morbi magna libero, interdum gravida ultricies a, vulputate vitae sem.",
     "publishedBy": "sam",
